refactor: separate funcs from server.ts
